Move PageDataOutput to presto-spi
